ALL DONE :D :D :D --- now, im just going to write some examples, then its bedtime. all this was done in like, 11 hours, plus i ate lunch, dinner, and went on a 30 min bike ride... thats what I call a WIN

+++ src/edu/wpi/tmathmeyer/security/network/packets/DynamicSizeOptionPacket.java
@@ -15,7 +15,7 @@ public class DynamicSizeOptionPacket implements Packet{
 	}
 	
 	public DynamicSizeOptionPacket(DataInputStream dis) throws IOException{
-		int size = dis.readByte();
+		int size = dis.readInt();
 		this.option = new byte[size];
 		for(int i=0;i<size;i++)
 			this.option[i] = dis.readByte();
